only delete password if mycnf is not empty

+++ nextcloud_update.sh
@@ -116,10 +116,13 @@ password='$regressionpw'
 LOGIN
     chmod 0600 $MYCNF
     chown root:root $MYCNF
-    rm /var/mysql_password.txt
     echo "Please restart the upgrade process, we fixed the password file $MYCNF."
     echo "Your password is:"
     cat "$MYCNF"
+    if [ ! -z "$MYSQLMYCNFPASS" ]
+    then
+        rm /var/mysql_password.txt
+    fi
     exit 1
 elif [ -z "$MYSQLMYCNFPASS" ] && [ -f /var/mysql_password.txt ]
 then
@@ -128,12 +131,18 @@ then
     echo "[client]"
     echo "password='$regressionpw'"
     } >> "$MYCNF"
-    rm /var/mysql_password.txt
+    if [ ! -z "$MYSQLMYCNFPASS" ]
+    then
+        rm /var/mysql_password.txt
+    fi
     echo "Please restart the upgrade process, we fixed the password file $MYCNF."
     echo "Your password is:"
     cat "$MYCNF"
     exit 1
 fi
+
+
+
 
 echo "Backing up files and upgrading to Nextcloud $NCVERSION in 10 seconds..."
 echo "Press CTRL+C to abort."
